Improvements especially to old tickets,
Be more careful overwriting keywords - keyword fields were overwriting
milestones for example. Perhaps the easiest thing to do would be to
maintain a set of projects so then Project.set could be used for
slightly nicer transaction history.

Milestone _|_ is mapped to ⊥.

If the owner is nobody it is removed. Sometimes the reporter is nobody
but you can't do anything about that.