Start with adding comment number infrastructure.
There needs to be a small rework as each transaction is actually a
"comment". So I should attach the number earlier in the pipeline, it is
not a huge deal losing some numbers from the map but the ordering has to
at least be right.

There is also the complication that multiple transactions can happen in
the same comment. We then have to choose which is the "primary"
transaction, the one which we want to assign the number. I think the way
to do this is to prefer comments and then it doesn't really matter.

There is also the problem that this means we must run the whole
migration in one go to maintain the CommentMap which is not recreatable
from the state of the database. This is not a good property to lose.